feat: add toolz to dependencies
Toolz is widely used in the codebase. In the past streamlit added toolz
as a transitive dependency, but that's no longer the case.

This change explicitly adds toolz to the various environment.yml files
to prevent import errors.

See also: 8beb1793 for how the perimeter-security-control app handles
this.

Should fix #21