refactor: unpin streamlit
The previous change (854a6c5) was done at a time when default version of
the streamlit library in Native Apps was 1.22. This app used some
features that required a fresher version than that.

As of BCR-1857 the default streamlit version is 1.35 which is enough for
the features the app is using.