refactor: move to use st.Page
Previously the code in `pages/` directory was very redundant. It was
essentially a wrapper around the `pipe(some_tiles, map(render))` code
which rendered said `some_tiles`.

This was in place because prior to release of st.Page in 1.38 the pages
in Streamlit application were tied to the presence of files.

Now the pages are constructed programmatically in a single python file.

I changed the ordering of the functions in the pipeline a bit so that
the sidebar completely rendered first before the tiles start rendering.

Prior to this change, the use of `st.connection` lead to the session
either not being created or `get_active_session` malfunctioning. I
worked around it using the try/except mechanism which is not ideal and
will need some further investigations.

Closes #17